# docs/swagger.yaml
basePath: /api
definitions:
  apperror.AppError:
    properties:
      code:
        type: integer
      error_code:
        description: Machine-readable error code
        type: string
      message:
        type: string
    type: object
...
  dto.LoginRequest:
    properties:
      password:
        example: password
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  dto.LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      message:
        example: Login successful
        type: string
      permissions:
        items:
          type: string
        type: array
      user:
        allOf:
        - $ref: '#/definitions/dto.UserResponse'
        description: Menggunakan UserResponse DTO
    type: object
...
  dto.RefreshTokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      avatar_url:
        example: https://example.com/avatar.png
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: a1b2c3d4-e5f6-7890-1234-567890abcdef
        type: string
      role:
        example: user
        type: string
      role_id:
        example: b1c2d3e4-f5g6-7890-1234-567890abcdef
        type: string
      username:
        example: johndoe
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API documentation for the Beresin App backend.
  title: Beresin App API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns an access token. The refresh token
        is set in an HttpOnly cookie.
      parameters:
      - description: Login Credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "400":
          description: Invalid request payload
          schema:
            $ref: '#/definitions/apperror.AppError'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/apperror.AppError'
      summary: User Login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidates the refresh token and clears the session cookie.
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Failed to read refresh token
          schema:
            $ref: '#/definitions/apperror.AppError'
      summary: Logout user
      tags:
      - Auth
  /auth/me:
    get:
      description: Returns the current authenticated user's information with their
        permissions.
      produces:
      - application/json
      responses:
        "200":
          description: Current user information
          schema:
            $ref: '#/definitions/dto.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/apperror.AppError'
      security:
      - BearerAuth: []
      summary: Get Current User Information
      tags:
      - Auth
  /auth/refresh:
    post:
      description: Generates a new access token using a valid refresh token from the
        cookie.
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/dto.RefreshTokenResponse'
        "401":
          description: Unauthorized or invalid refresh token
          schema:
            $ref: '#/definitions/apperror.AppError'
      summary: Refresh Access Token
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"